+++ src/student_aleksandrs_jakovenko/Lesson2/Level2/BankAccount.java
@@ -1,0 +1,7 @@
+package student_aleksandrs_jakovenko.Lesson2.Level2;
+
+public class BankAccount {
+    public static void main(String[] args) {
+
+    }
+}
