+++ src/student_aleksandrs_jakovenko/Lesson2/Level2/LoginService.java
@@ -1,0 +1,7 @@
+package student_aleksandrs_jakovenko.Lesson2.Level2;
+
+public class LoginService {
+    public static void main(String[] args) {
+        System.out.println("You can login!");
+    }
+}
