+++ src/student_aleksandrs_jakovenko/Lesson2/Level2/Phone.java
@@ -1,0 +1,7 @@
+package student_aleksandrs_jakovenko.Lesson2.Level2;
+
+public class Phone {
+    public static void main(String[] args) {
+
+    }
+}
