+++ src/student_aleksandrs_jakovenko/Lesson2/Level2/Task8.java
@@ -1,0 +1,13 @@
+package student_aleksandrs_jakovenko.Lesson2.Level2;
+
+import java.util.Scanner;
+
+public class Task8 {
+    public static void main(String[] args) {
+        Scanner scanner = new Scanner(System.in);
+
+        System.out.println("Enter your name:");
+        String name = scanner.nextLine();
+        System.out.println("Hello " + name + "!");
+    }
+}
