+++ src/student_aleksandrs_jakovenko/Lesson2/Level2/Turtle.java
@@ -1,0 +1,7 @@
+package student_aleksandrs_jakovenko.Lesson2.Level2;
+
+public class Turtle {
+    public static void main(String[] args) {
+
+    }
+}
